Initial Commit
Building on basics from RPGPoker project (which was in this account a the time it was initally commited).
Adding a database so data persists and can be reviewed - and everything that goes with creating and analyzing persistant data.

This isn't a clean add.  Some files aren't currently used but were  used during development.