Teams removal, damage and XP
I removed some of the code originally in place to start with teams instead of Characters.
I inserted a list into the damage function to start the work of letting the user define it.
I realized the Databasebattleupdate never updated the character table with XP, defeats, survivals and added that.  Though it insn't totally working as expected yet.